refactor(client): splitted model logic from source logic
